+++ src/meson.build
@@ -1,2 +1,3 @@
 subdir ('brc')
+subdir ('Handlers')
 subdir ('lib')
