+++ test/meson.build
@@ -1,4 +1,4 @@
-executable('fake-enc', 'fake-enc-test.c',
+executable('fake-enc', 'fake-enc-test.cpp',
   include_directories: libmebo_inc,
-  dependencies: libmebo_dep_internal,
+  dependencies: [libmebo_dep_internal, libbrc_controller_dep],
   install: false)
